+++ src/main/java/wooteco/subway/line/domain/repository/SectionRepository.java
@@ -1,8 +1,9 @@
-package wooteco.subway.line.repository;
+package wooteco.subway.line.domain.repository;
 
 import org.springframework.stereotype.Repository;
 import wooteco.subway.line.domain.section.Section;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
@@ -10,4 +11,8 @@ public interface SectionRepository {
     Section save(Section section);
 
     Optional<Section> findById(Long id);
+
+    List<Section> findByLineId(Long lineId);
+
+    void remove(Long id);
 }
