+++ src/main/java/wooteco/subway/line/infra/JdbcSectionRepository.java
@@ -1,0 +1,69 @@
+package wooteco.subway.line.infra;
+
+import org.springframework.dao.DataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+import wooteco.subway.line.domain.repository.SectionRepository;
+import wooteco.subway.line.domain.section.Distance;
+import wooteco.subway.line.domain.section.Section;
+
+import java.sql.PreparedStatement;
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public class JdbcSectionRepository implements SectionRepository {
+    private final JdbcTemplate jdbcTemplate;
+    private final RowMapper<Section> sectionRowMapper = (resultSet, rowNum) ->
+            new Section(
+                    resultSet.getLong("id"),
+                    resultSet.getLong("line_id"),
+                    resultSet.getLong("up_station_id"),
+                    resultSet.getLong("down_station_id"),
+                    new Distance(resultSet.getInt("distance"))
+            );
+
+    public JdbcSectionRepository(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public Section save(Section section) {
+        String query = "INSERT INTO section(line_id, up_station_id, down_station_id, distance) VALUES(?,?,?,?)";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        this.jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(query, new String[]{"id"});
+            ps.setLong(1, section.getLineId());
+            ps.setLong(2, section.getUpStationId());
+            ps.setLong(3, section.getDownStationId());
+            ps.setInt(4, section.getDistance().value());
+            return ps;
+        }, keyHolder);
+        return findById(keyHolder.getKey().longValue()).get();
+    }
+
+    @Override
+    public Optional<Section> findById(Long id) {
+        try {
+            String query = "SELECT * FROM section WHERE id = ?";
+            return Optional.ofNullable(jdbcTemplate.queryForObject(query, sectionRowMapper, id));
+        } catch (DataAccessException e) {
+            return Optional.empty();
+        }
+    }
+
+    @Override
+    public List<Section> findByLineId(Long lineId) {
+        String query = "SELECT * FROM section WHERE line_id = ?";
+        return jdbcTemplate.query(query, sectionRowMapper, lineId);
+    }
+
+    @Override
+    public void remove(Long id) {
+        String query = "DELETE FROM section WHERE id = ?";
+        jdbcTemplate.update(query, id);
+    }
+}
